Use same field mapping for solrsearch and listing endpoints.
